tekton: update cel expressions to avoid build loops
- When bundle files change, build the bundle.
- Otherwise, build the operator.

The components relationship on Konflux are configured so that,
when the operator changes, a successful operator build will trigger
a pull request to update the image references in the bundle.
That in turn will trigger a bundle build.
See https://konflux-ci.dev/docs/how-tos/configuring/component-nudges/.

We should avoid changing both the operator and the bundle at the same time.